Divided unit test file into smaller test files
A directory has been created for invoices and processors.
`unit_test.py` has been removed. This led to a small update
in `workflows/unit-test.yaml`
