+++ process_report/tests/unit/invoices/test_pi_specific_invoice.py
@@ -1,0 +1,68 @@
+from unittest import TestCase, mock
+import tempfile
+import pandas
+import os
+
+from process_report.tests import util as test_utils
+
+
+class TestExportPICSV(TestCase):
+    def setUp(self):
+        data = {
+            "Invoice Month": ["2023-01", "2023-01", "2023-01", "2023-01", "2023-01"],
+            "Manager (PI)": ["PI1", "PI1", "PI1", "PI2", "PI2"],
+            "Institution": ["BU", "BU", "BU", "HU", "HU"],
+            "Project - Allocation": [
+                "ProjectA",
+                "ProjectB",
+                "ProjectC",
+                "ProjectD",
+                "ProjectE",
+            ],
+            "Untouch Data Column": ["DataA", "DataB", "DataC", "DataD", "DataE"],
+            "Is Billable": [True, True, True, True, True],
+            "Missing PI": [False, False, False, False, False],
+        }
+        self.dataframe = pandas.DataFrame(data)
+        self.invoice_month = data["Invoice Month"][0]
+
+    @mock.patch("process_report.invoices.invoice.Invoice._filter_columns")
+    def test_export_pi(self, mock_filter_cols):
+        mock_filter_cols.return_value = self.dataframe
+
+        output_dir = tempfile.TemporaryDirectory()
+        pi_inv = test_utils.new_pi_specific_invoice(
+            output_dir.name, invoice_month=self.invoice_month, data=self.dataframe
+        )
+        pi_inv.process()
+        pi_inv.export()
+        pi_csv_1 = f'{self.dataframe["Institution"][0]}_{self.dataframe["Manager (PI)"][0]} {self.dataframe["Invoice Month"][0]}.csv'
+        pi_csv_2 = f'{self.dataframe["Institution"][3]}_{self.dataframe["Manager (PI)"][3]} {self.dataframe["Invoice Month"][3]}.csv'
+        self.assertIn(pi_csv_1, os.listdir(output_dir.name))
+        self.assertIn(pi_csv_2, os.listdir(output_dir.name))
+        self.assertEqual(
+            len(os.listdir(output_dir.name)),
+            len(self.dataframe["Manager (PI)"].unique()),
+        )
+
+        pi_df = pandas.read_csv(output_dir.name + "/" + pi_csv_1)
+        self.assertEqual(len(pi_df["Manager (PI)"].unique()), 1)
+        self.assertEqual(
+            pi_df["Manager (PI)"].unique()[0], self.dataframe["Manager (PI)"][0]
+        )
+
+        self.assertIn("ProjectA", pi_df["Project - Allocation"].tolist())
+        self.assertIn("ProjectB", pi_df["Project - Allocation"].tolist())
+        self.assertIn("ProjectC", pi_df["Project - Allocation"].tolist())
+
+        pi_df = pandas.read_csv(output_dir.name + "/" + pi_csv_2)
+        self.assertEqual(len(pi_df["Manager (PI)"].unique()), 1)
+        self.assertEqual(
+            pi_df["Manager (PI)"].unique()[0], self.dataframe["Manager (PI)"][3]
+        )
+
+        self.assertIn("ProjectD", pi_df["Project - Allocation"].tolist())
+        self.assertIn("ProjectE", pi_df["Project - Allocation"].tolist())
+        self.assertNotIn("ProjectA", pi_df["Project - Allocation"].tolist())
+        self.assertNotIn("ProjectB", pi_df["Project - Allocation"].tolist())
+        self.assertNotIn("ProjectC", pi_df["Project - Allocation"].tolist())
