+++ process_report/tests/unit/processors/test_validate_billable_pi_processor.py
@@ -1,0 +1,52 @@
+from unittest import TestCase
+import pandas
+import uuid
+import math
+
+from process_report.tests import util as test_utils
+
+
+class TestValidateBillablePIProcessor(TestCase):
+    def test_remove_nonbillables(self):
+        pis = [uuid.uuid4().hex for x in range(10)]
+        projects = [uuid.uuid4().hex for x in range(10)]
+        nonbillable_pis = pis[:3]
+        nonbillable_projects = projects[7:]
+        billable_pis = pis[3:7]
+        data = pandas.DataFrame({"Manager (PI)": pis, "Project - Allocation": projects})
+
+        validate_billable_pi_proc = test_utils.new_validate_billable_pi_processor(
+            data=data,
+            nonbillable_pis=nonbillable_pis,
+            nonbillable_projects=nonbillable_projects,
+        )
+        validate_billable_pi_proc.process()
+        data = validate_billable_pi_proc.data
+        data = data[data["Is Billable"]]
+        self.assertTrue(data[data["Manager (PI)"].isin(nonbillable_pis)].empty)
+        self.assertTrue(
+            data[data["Project - Allocation"].isin(nonbillable_projects)].empty
+        )
+        self.assertTrue(data["Manager (PI)"].isin(billable_pis).all())
+
+    def test_empty_pi_name(self):
+        test_data = pandas.DataFrame(
+            {
+                "Manager (PI)": ["PI1", math.nan, "PI1", "PI2", "PI2"],
+                "Project - Allocation": [
+                    "ProjectA",
+                    "ProjectB",
+                    "ProjectC",
+                    "ProjectD",
+                    "ProjectE",
+                ],
+            }
+        )
+        self.assertEqual(1, len(test_data[pandas.isna(test_data["Manager (PI)"])]))
+        validate_billable_pi_proc = test_utils.new_validate_billable_pi_processor(
+            data=test_data
+        )
+        validate_billable_pi_proc.process()
+        output_data = validate_billable_pi_proc.data
+        output_data = output_data[~output_data["Missing PI"]]
+        self.assertEqual(0, len(output_data[pandas.isna(output_data["Manager (PI)"])]))
